requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
python-docx
Pillow
python-multipart